Change specialized and core docs to inline docstrings
