+++ src/CUTEst.jl
@@ -65,7 +65,6 @@ include("core_interface.jl")
 include("specialized_interface.jl")
 include("julia_interface.jl")
 include("mpb_interface.jl")
-include("documentation.jl")
 
 """Decode problem and build shared library.
